+++ src/labkey/Domain.ts
@@ -408,6 +408,30 @@ export function save(config: SaveDomainOptions): XMLHttpRequest {
     });
 }
 
+/**
+ * Validate name expression(s) for domain design.
+ */
+export function validateNameExpressions(config: SaveDomainOptions): XMLHttpRequest {
+
+    let options: SaveDomainOptions = arguments.length > 1 ? {
+        success: arguments[0],
+        failure: arguments[1],
+        domainDesign: arguments[2],
+        options: arguments[3],
+    } : config;
+
+    return request({
+        url: buildURL('property', 'validateNameExpressions.api', options.containerPath),
+        method: 'GET',
+        success: getCallbackWrapper(getOnSuccess(options), options.scope),
+        failure: getCallbackWrapper(getOnFailure(options), options.scope, true),
+        jsonData: {
+            domainDesign: options.domainDesign,
+            options: options.options,
+        }
+    });
+}
+
 export interface UpdateDomainParams {
     createFields?: []
     deleteFields?: number[]